requirements.txt
rich
Pillow